requirements.txt
rich
Pillow